refactor: 바인딩 어댑터 파일 분리
