+++ app/src/main/java/woowacourse/movie/presentation/bindingadapter/ImageViewBindingAdapter.kt
@@ -1,0 +1,13 @@
+package woowacourse.movie.presentation.bindingadapter
+
+import android.widget.ImageView
+import androidx.annotation.DrawableRes
+import androidx.databinding.BindingAdapter
+
+@BindingAdapter("imgResPath")
+fun setImageViewResourcePath(
+    view: ImageView,
+    @DrawableRes resId: Int,
+) {
+    view.setImageResource(resId)
+}
